+++ CHANGELOG.md
@@ -5,6 +5,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed nxService causing an unknown propertyName error.
+
+## [1.1.0] - 2023-07-18
+
 ### Added
 
 - DscResource: